requirements.txt
asgiref==3.2.7
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
Django==3.0.7
django-cors-headers==3.2.1
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
Jinja2==2.11.2
Markdown==3.2.1
psycopg2==2.8.5
py==1.8.1
pytest==5.4.1
pytest-cov==2.8.1
pytest-django==3.9.0
pytest-pythonpath==0.7.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.9

